Move stream transmit logic into Streams
