+++ quinn-proto/src/streams.rs
@@ -1,12 +1,19 @@
-use std::collections::{hash_map, HashMap};
+use std::{
+    collections::{hash_map, HashMap},
+    mem,
+};
 
 use bytes::Bytes;
 use err_derive::Error;
 use tracing::debug;
 
 use crate::{
-    assembler::Assembler, frame, range_set::RangeSet, send_buffer::SendBuffer,
-    transport_parameters::TransportParameters, Dir, Side, StreamId, TransportError, VarInt,
+    assembler::Assembler,
+    frame::{self, FrameStruct},
+    range_set::RangeSet,
+    send_buffer::SendBuffer,
+    transport_parameters::TransportParameters,
+    Dir, Side, StreamId, TransportError, VarInt,
 };
 
 pub(crate) struct Streams {
@@ -237,6 +244,34 @@ impl Streams {
 
     pub fn can_send(&self) -> bool {
         !self.pending.is_empty()
+    }
+
+    /// Get data to send on a stream frame, if any is available
+    pub fn poll_transmit(&mut self, max_frame_size: usize) -> Option<frame::StreamMeta> {
+        let max_data_len = max_frame_size.checked_sub(frame::Stream::SIZE_BOUND)?;
+        loop {
+            let id = self.pending.pop()?;
+            let stream = match self.send.get_mut(&id) {
+                Some(s) => s,
+                // Stream was reset with pending data and the reset was acknowledged
+                None => continue,
+            };
+            // Reset streams aren't removed from the pending list and still exist while the peer
+            // hasn't acknowledged the reset, but should not generate STREAM frames, so we need to
+            // check for them explicitly.
+            if stream.is_reset() {
+                continue;
+            }
+            let offsets = stream.pending.poll_transmit(max_data_len);
+            let fin = offsets.end == stream.pending.offset()
+                && mem::replace(&mut stream.fin_pending, false);
+            if stream.is_pending() {
+                self.pending.push(id);
+            }
+            // Would be nice to return a slice directly here as well so the caller doesn't have to
+            // call `pending_data` and redo the hash lookup, but borrowck objects.
+            return Some(frame::StreamMeta { id, offsets, fin });
+        }
     }
 
     /// Returns whether the stream was finished
